Add dispatch database schema
This is designed to allow cloud-side decision-makers to request dispatch
of resources from individual edge locations.

Dispatch is described in terms of the type of resource to be activated,
in a specific location, over a specific time interval.  The status and
settings of this dispatch may change many times, but this triple setting
of `(location_id, resource_id, time_interval_id)` forms a unique ID for
the dispatch entity itself.

The schema is designed to be purely additive, i.e. once created each row
can/should be treated as immutable.  Dispatch status and settings can be
updated by writing a new `dispatch_modification` row, the most recent of
which (by `create_time`) determines the current values.

`location_id` values are treated purely as numbers in this schema: it is
assumed that they are defined (and linked e.g. to enterprises, TSO/DSOs,
etc.) in some other database.